# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hangman
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
paths:
  /:
    x-swagger-router-controller: home
    post:
      description: Posts the letter selected by the user and triggers side effects
      operationId: pickLetter
      parameters:
      - name: guess
        in: body
        schema:
          type: string
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/Game"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Displays game interface
      operationId: game
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Game"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Game:
    description: The game object
    type: object
    properties:
      secretWordChars:
        # how can I reuse this type?
        type: array
        items:
          type: object
          properties:
            value:
              type: string
            isGuessed:
              type: boolean
      letters:
        type: array
        items:
          type: object
          properties:
            value:
              type: string
            isGuessed:
              type: boolean
      guessesRemaining:
        type: integer
      won:
        type: boolean
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string